+++ std-bits/table/src/main/java/org/enso/table/data/column/operation/map/numeric/DoubleLongMapOpWithSpecialNumericHandling.java
@@ -1,6 +1,9 @@
-package org.enso.table.data.column.operation.map;
+package org.enso.table.data.column.operation.map.numeric;
 
 import java.util.BitSet;
+
+import org.enso.table.data.column.operation.map.MapOperationProblemBuilder;
+import org.enso.table.data.column.operation.map.UnaryMapOperationWithProblemBuilder;
 import org.enso.table.data.column.storage.numeric.DoubleStorage;
 import org.enso.table.data.column.storage.numeric.LongStorage;
 import org.graalvm.polyglot.Context;
