+++ std-bits/table/src/main/java/org/enso/table/data/column/operation/map/numeric/UnaryDoubleToLongOp.java
@@ -1,6 +1,8 @@
-package org.enso.table.data.column.operation.map;
+package org.enso.table.data.column.operation.map.numeric;
 
 import java.util.BitSet;
+
+import org.enso.table.data.column.operation.map.UnaryMapOperation;
 import org.enso.table.data.column.storage.numeric.DoubleStorage;
 import org.enso.table.data.column.storage.numeric.LongStorage;
 import org.graalvm.polyglot.Context;
