+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/datetime/DateStorage.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 import org.enso.table.data.column.builder.Builder;
 import org.enso.table.data.column.builder.DateBuilder;
 import org.enso.table.data.column.operation.map.MapOpStorage;
-import org.enso.table.data.column.operation.map.UnaryIntegerOp;
+import org.enso.table.data.column.operation.map.numeric.UnaryIntegerOp;
 import org.enso.table.data.column.operation.map.datetime.DateTimeIsInOp;
 import org.enso.table.data.column.storage.ObjectStorage;
 import org.enso.table.data.column.storage.SpecializedStorage;
